Generalize file downloading (closes #9)
