+++ lib/bdc-availability/download/download.js
@@ -1,12 +1,9 @@
 import {
-  createWriteStream as fsCreateWriteStream,
   existsSync as fsExistsSync,
   mkdirSync as fsMkdirSync,
 } from 'fs';
 import { resolve as pathResolve } from 'path';
-import { Readable as StreamReadable } from 'stream';
-import { finished as streamFinished } from 'stream/promises';
-import { mapLimit as asyncMapLimit } from 'async';
+import downloadFiles from '../../utils/download-files.js';
 import {
   resolveTilde,
   warn,
@@ -116,40 +113,6 @@ async function fetchFilingProcessId(targetFilingDateString) {
   }
 
   return filingProcessId;
-}
-
-async function downloadBdcFile(saveDirPath, url, fileName) {
-  // TODO better progress reporting
-  console.log('Downloading', url);
-
-  // note: default behavior is to overwrite the file
-  const outPath = `${saveDirPath}/${fileName}.zip`
-  const writeStream = fsCreateWriteStream(outPath);  
-  
-  const { body } = await fetch(url);
-
-  // convert web stream to readable stream
-  const readable = StreamReadable.fromWeb(body);
-
-  return streamFinished(readable.pipe(writeStream));
-
-  // TODO don't do this if they pass in the future "don't unzip" flag
-  // await unzipFile(outPath, saveDirPath);
-}
-
-async function downloadBdcFiles(
-  saveDirPath,
-  urlsAndFilesNames,
-  numConcurrentRequests = 3,
-) {
-  return asyncMapLimit(
-    urlsAndFilesNames,
-    numConcurrentRequests,
-    async (urlAndFileName) => {
-      const { url, fileName } = urlAndFileName;
-      return downloadBdcFile(saveDirPath, url, fileName);
-    },
-  );
 }
 
 
@@ -281,18 +244,18 @@ async function download(options) {
     };
   });
 
-  await downloadBdcFiles(outDirPath, urlsAndFileNames);
+  await downloadFiles(outDirPath, urlsAndFileNames);
 } 
 
 // DEBUG
-// download({
-//   entityType: 'states',
-//   entityIds: 'ri',
-//   filingType: 'fixed',
-//   outDir: '~/Downloads/bbkit-test',
-//   // tech: 'fiber',
-//   // format: 'shp',
-//   // filingDate: '',
-// });
+download({
+  entityType: 'states',
+  entityIds: 'ri',
+  filingType: 'fixed',
+  outDir: '~/Downloads/bbkit-test',
+  // tech: 'fiber',
+  // format: 'shp',
+  // filingDate: '',
+});
 
 export default download;
